requirements/_base.txt
asgiref==3.2.10
beautifulsoup4==4.8.2
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.1.2
django-axes==5.8.0
django-filter==2.4.0
django-ipware==3.0.1
django-modelcluster==5.1
django-taggit==1.3.0
django-treebeard==4.3.1
django-weasyprint==1.0.2
djangorestframework==3.12.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2018.5
openpyxl==3.0.5
Pillow==7.2.0
pycparser==2.20
Pyphen==0.9.5
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
tinycss2==1.0.2
Unidecode==1.1.1
urllib3==1.25.11
wagtail==2.10.2
WeasyPrint==51
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
